requirements.txt
Django
djangorestframework
drf-nested-routers
